requirements.txt
# Core agentic framework
crewai[tools]>=0.157.0,<1.0.0

# API server
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# GitHub integration
PyGithub>=2.5.0

# RAG - vector store + Ollama embeddings
chromadb>=0.5.0

# HTTP client (used by RAG tool to call Ollama embeddings API)
requests>=2.32.0
